Removed chargemon executables, these where still included but are really, really old...
Changed the code for the hijack backup of the stock binary. It will check if the stock file is there and if the chargemon file in system/bin is a binary or a script before moving it away to the backup file.

Removed some smaller redundant stuff, also tried to improve the failed installation detection.

+++ src/backupstockbinaries.sh
@@ -6,28 +6,8 @@ if [ ! -f "/system/bin/mr.stock" ]; then
 
 fi
 
-if [ ! -f "/system/bin/chargemon.stock" ]; then
+if [ ! -f "/system/bin/chargemon.stock" -a "$(/sbin/busybox head -n 1 /system/bin/chargemon)" != '#!/system/bin/sh' ]; then
 
 	/sbin/busybox mv /system/bin/chargemon /system/bin/chargemon.stock
-
-fi
-
-# Just to make sure it's not a script
-CHARGEMON=`/sbin/busybox head -n 1 /system/bin/chargemon.stock`
-
-if [ "${CHARGEMON}" = "#!/system/bin/sh" ]; then
-
-	# It is, lets replace it with the correct binary!
-	MODEL=`/system/bin/getprop ro.product.model`
-	if [ "${MODEL}" = "C6602" -o "${MODEL}" = "C6603" -o "${MODEL}" = "SO-02E" ]; then
-		/sbin/busybox cp /tmp/chargemon.c660x /system/bin/chargemon.stock
-		/sbin/busybox chmod 755 /system/bin/chargemon.stock
-	elif [ "${MODEL}" = "C6502" -o "${MODEL}" = "C6503" -o "${MODEL}" = "C6506" ]; then
-		/sbin/busybox cp /tmp/chargemon.c650x /system/bin/chargemon.stock
-		/sbin/busybox chmod 755 /system/bin/chargemon.stock
-	elif [ "${MODEL}" = "C6902" -o "${MODEL}" = "C6903" -o "${MODEL}" = "C6906" -o "${MODEL}" = "C6943" ]; then
-		/sbin/busybox cp /tmp/chargemon.c690x /system/bin/chargemon.stock
-		/sbin/busybox chmod 755 /system/bin/chargemon.stock
-	fi
 
 fi
